+++ receiver/windowsperfcountersreceiver/config_test.go
@@ -115,7 +115,7 @@ func TestLoadConfig_Error(t *testing.T) {
 			name:    "AllErrors",
 			cfgFile: "config-allerrors.yaml",
 			expectedErr: fmt.Sprintf(
-				"%s: [%s; %s; %s; %s]",
+				"%s: %s; %s; %s; %s",
 				errorPrefix,
 				negativeCollectionIntervalErr,
 				fmt.Sprintf(emptyInstanceErr, "object"),